requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
click==8.1.7
    # via typer
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pygments==2.17.2
    # via rich
pypandoc==1.13
    # via -r requirements.in
python-frontmatter==1.1.0
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pyyaml==6.0.1
    # via python-frontmatter
rich==13.7.1
    # via
    #   -r requirements.in
    #   typer
shellingham==1.5.4
    # via typer
text-unidecode==1.3
    # via python-slugify
typer==0.12.3
    # via -r requirements.in
typing-extensions==4.11.0
    # via typer
